🚚 refactor : move ApiKey to config folder
