+++ src/main/java/com/dodatabase/movie_backend/config/WebClientConfig.java
@@ -4,8 +4,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import com.dodatabase.movie_backend.service.ApiKey;
-
 import lombok.Builder;
 
 @Configuration
